Move queries into hooks to prevent prop drilling
